Vitae.ai-backend/backend/requirements.txt
flask
flask-restful
sqlalchemy
PyPDF2
openai
python-dotenv
